requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Requests==2.31.0
gunicorn==21.2.0
threaded==4.1.0
psycopg2==2.9.7
Flask-Session==0.5.0